+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Apr 2018
+
+- Move bash scripts setup to belt js commands
+
 ## Dec 2017
 
 - Switched Atom with VS Code as a main editor. Faster but has some quirks. Still use Atom for my wiki